Query: Merge AsyncQueryingEnumerable into QueryingEnumerable
It is just a wrapper to create enumerators. Enumerators can be created out of single Enumerable
